Reorganize file structure
and adds tests GenericInitailizer
